breaking up filters duck
