+++ src/components/Filters/FilterTristate.js
@@ -2,7 +2,8 @@
 import React, { memo } from "react";
 import TristateCheckbox from "components/Filters/TristateCheckbox";
 import { useSelector, useDispatch } from "react-redux";
-import { selectFilterAtIndex, updateFilterTristate } from "redux-ducks/filters";
+import { selectFilterAtIndex } from "redux-ducks/filters";
+import { updateFilterTristate } from "redux-ducks/filters/actionCreators";
 
 type Props = { index: number };
